+++ packages/yourloops/test/integration/mock/patient.api.mock.ts
@@ -121,7 +121,7 @@ export const monitoredPatient: Patient = buildPatient(
     },
     {
       teamId: myThirdTeamId,
-      status: UserInvitationStatus.accepted,
+      status: UserInvitationStatus.accepted
     }
   ],
   defaultMonitoring,